backend/foodgram/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
decouple==0.0.7
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.4.0
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.2.0
drf-base64==2.0
drf-spectacular==0.27.2
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
